reimbursement-management/backend/requirements.txt
fastapi==0.115.12
uvicorn[standard]==0.34.0
sqlalchemy==2.0.40
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-multipart==0.0.20
httpx==0.28.1
pillow==11.1.0
pytesseract==0.3.13
aiofiles==24.1.0
python-dotenv==1.1.0
alembic==1.15.2
pydantic[email]==2.11.1
